Factored out session utility functions into a helper
